cleanup
Cleanup files associated with the tkInter GUI
